fix some bugs

+++ cc-im/src/main/java/com/zj/ccIm/core/api/FunctionApi.kt
@@ -3,7 +3,6 @@ package com.zj.ccIm.core.api
 import com.zj.ccIm.core.bean.FetchPrivateOwnerSessionBean
 import com.zj.database.entity.SessionLastMsgInfo
 import com.zj.ccIm.core.bean.FetcherSessionBean
-import com.zj.database.entity.MessageInfoEntity
 import io.reactivex.Observable
 import retrofit2.http.*
 
@@ -22,5 +21,5 @@ interface FunctionApi {
 
     @Multipart
     @POST("/im/group/private/chat/remove")
-    fun deleteSession(@Field("targetUserId") targetUserId: Long?, @Field("groupId") groupId: Long?, @Field("status") status: Int?): Observable<String?>
+    fun deleteSession(@Field("targetUserId") targetUserId: Int?, @Field("groupId") groupId: Long?, @Field("status") status: Int?): Observable<String?>
 }
