+++ cc-im/src/main/java/com/zj/ccIm/core/bean/GetMsgReqBean.kt
@@ -3,7 +3,7 @@ package com.zj.ccIm.core.bean
 import com.zj.ccIm.annos.MsgFetchType
 import com.zj.ccIm.core.fecher.FetchMsgChannel
 
-data class GetMsgReqBean constructor(val groupId: Long, val ownerId: Long, val targetUserid: Long?, val msgId: Long?, @MsgFetchType val type: Int? = null, var channels: Array<out FetchMsgChannel>) {
+data class GetMsgReqBean constructor(val groupId: Long, val ownerId: Int, val targetUserid: Int?, val msgId: Long?, @MsgFetchType val type: Int? = null, var channels: Array<out FetchMsgChannel>) {
 
     private var channel: List<String>? = null
